+++ src/shared.js
@@ -1,50 +1,4 @@
 /**
- * Special attributes that map to DOM events.
- */
-
-export let events = {
-  onBlur: 'blur',
-  onChange: 'change',
-  onClick: 'click',
-  onContextMenu: 'contextmenu',
-  onCopy: 'copy',
-  onCut: 'cut',
-  onDoubleClick: 'dblclick',
-  onDrag: 'drag',
-  onDragEnd: 'dragend',
-  onDragEnter: 'dragenter',
-  onDragExit: 'dragexit',
-  onDragLeave: 'dragleave',
-  onDragOver: 'dragover',
-  onDragStart: 'dragstart',
-  onDrop: 'drop',
-  onError: 'error',
-  onFocus: 'focus',
-  onInput: 'input',
-  onInvalid: 'invalid',
-  onKeyDown: 'keydown',
-  onKeyPress: 'keypress',
-  onKeyUp: 'keyup',
-  onLoad: 'load',
-  onMouseDown: 'mousedown',
-  onMouseEnter: 'mouseenter',
-  onMouseLeave: 'mouseleave',
-  onMouseMove: 'mousemove',
-  onMouseOut: 'mouseout',
-  onMouseOver: 'mouseover',
-  onMouseUp: 'mouseup',
-  onPaste: 'paste',
-  onReset: 'reset',
-  onScroll: 'scroll',
-  onSubmit: 'submit',
-  onTouchCancel: 'touchcancel',
-  onTouchEnd: 'touchend',
-  onTouchMove: 'touchmove',
-  onTouchStart: 'touchstart',
-  onWheel: 'wheel'
-}
-
-/**
  * Determine if two virtual nodes are the same type
  */
 
@@ -56,7 +10,7 @@ export let isSameType = (left, right) => {
  * Get the content for a custom element
  */
 
-export let renderCustomElement = (customElement, model, context) => {
+export let renderThunk = (customElement, model, context) => {
   let render = customElement.type
   let rootElement = render(model, context)
   return rootElement
@@ -66,11 +20,12 @@ export let renderCustomElement = (customElement, model, context) => {
  * The model is used in the custom element render function.
  */
 
-export let createModel = (virtualElement, path) => {
+export let createModel = (virtualElement, context, path) => {
   return {
     attributes: virtualElement.attributes,
     children: virtualElement.children,
-    path: path
+    path: path,
+    context: context
   }
 }
 
@@ -100,19 +55,6 @@ export let getKey = (virtualElement) => {
 }
 
 /**
- * Check if an attribute shoudl be rendered into the DOM.
- */
-
-export let isActiveAttribute = (value) => {
-  if (typeof value === 'boolean') return value
-  if (value === '') return true
-  if (value === undefined) return false
-  if (value === null) return false
-  if (value === false) return false
-  return true
-}
-
-/**
  * Get the type of virtual element.
  */
 
@@ -121,7 +63,7 @@ export let nodeType = (element) => {
     case 'string':
       return 'native'
     case 'function':
-      return 'custom'
+      return 'thunk'
     default:
       return 'text'
   }
