+++ src/svg.js
@@ -1,0 +1,10 @@
+import {isElement} from 'is-svg-element'
+import isAttribute from 'is-svg-attribute'
+
+const namespace = 'http://www.w3.org/2000/svg'
+
+export default {
+  isElement,
+  isAttribute,
+  namespace
+}
